switch to debug level
