+++ tests/test_utils.py
@@ -9,8 +9,6 @@ import pytest
 from mockito import mock, when
 from requests.exceptions import HTTPError
 from urllib3.exceptions import MaxRetryError
-import logging
-from terralab.utils import RetryMessageFilter
 from terralab import utils
 from terralab.utils import handle_api_exceptions
 from tests.utils_for_tests import capture_logs
@@ -200,50 +198,6 @@ def test_server_unavailable_error(capture_logs):
         "Maximum retries reached. The server may be down or unreachable. Please try again later."
         in capture_logs.text
     )
-
-
-def test_retry_message_filter_match():
-    filter_instance = RetryMessageFilter()
-
-    record = logging.LogRecord(
-        name="urllib3.connectionpool",
-        level=logging.WARNING,
-        pathname="test_path",
-        lineno=10,
-        msg="Retrying (Retry(total=2, connect=None, read=None, redirect=None, status=None)) "
-        "after connection broken by 'NewConnectionError': /api/test",
-        args=(),
-        exc_info=None,
-    )
-
-    result = filter_instance.filter(record)
-
-    assert result is True
-    assert (
-        record.msg
-        == "terralab encountered a problem connecting to the server. Retrying your request..."
-    )
-    assert record.args == ()
-
-
-# Tests that we don't clobber other non-retry log messages
-def test_retry_message_filter_non_match():
-    filter_instance = RetryMessageFilter()
-
-    non_matching_record = logging.LogRecord(
-        name="urllib3.connectionpool",
-        level=logging.WARNING,
-        pathname="test_path",
-        lineno=10,
-        msg="Some other log message",
-        args=(),
-        exc_info=None,
-    )
-
-    result = filter_instance.filter(non_matching_record)
-
-    assert result is True
-    assert non_matching_record.msg == "Some other log message"
 
 
 format_timestamp_testdata = [
